+++ app/Filament/Widgets/Activity/Charts/DurationPerDayChart.php
@@ -29,7 +29,7 @@ class DurationPerDayChart extends ChartWidget
         $fromDay = $from->copy()->startOfDay();
         $toDay = $to->copy()->startOfDay();
 
-        $days = max($fromDay->diffInDays($toDay) + 1, 1);
+        $days = (int)max($fromDay->diffInDays($toDay) + 1, 1);
 
         $q = Activity::query()
             ->whereBetween("created_at", [$from, $to]);
